somarim-cost-optimized/requirements.txt
# requirements.txt
aws-cdk-lib>=2.80.0
constructs>=10.0.0
